We do a little fixing >:) i hate pyinstaller
Why does Pyinstaller automatically default to _internal? Why does it exist? Why does it need to exist? Pyinstaller is ruining my perfectly fine code. It takes so long to make, only to find a bug that exists ONLY IN PYINSTALLER. ASFKDJHAEIRKJAKLGJKASDJFKSDLFJKASJKLVELC
Oh hey, you're still here. The fix was just to chdir .. if you are in internal.